+++ server.js
@@ -14,6 +14,7 @@ app.use(bodyParser.urlencoded({ extended: true }));
 
 require('./app/index');
 require('./app/routes/drinks.routes')(app);
+require('./app/routes/reviews.routes')(app);
 require('./app/routes/users.routes')(app);
 
 const PORT = process.env.PORT || 8080
